requirements.txt
openai-whisper>=20231117
transformers>=4.38.0
torch>=2.1.0
torchaudio>=2.1.0
soundfile>=0.12.1
numpy>=1.24.0
streamlit>=1.32.0
datasets>=2.18.0
jiwer>=3.0.3
pytest>=8.0.0
